Kitchen Stocked
Kitchen Boundaries are established and ingredients are able to be collected by the players. The players are able to hold two ingredients at a time.

Next Goal: Add the Chopping Stations where the player will be able to chop their veggies before adding them to a bowl. They will also be able to use the trash can to toss items in their main hand as well as bowls they are holding.