The First Customer
Instead of developing the Chopping Stations, I instead added the First Customer. The Customer spawns in the middle first and can have an order that contains up to 3 ingredients. Orders after the first will fill in from the sides to offer a balanced chance for both players to get to customers.

Next Goal: Chopping Station, trash can, and delivering an order to a customer.